legalease/requirements.txt
streamlit
langchain
sentence-transformers
groq
PyMuPDF
python-dotenv
faiss-cpu
